+++ modules/common/ghostship-pkg.nix
@@ -1,6 +1,4 @@
 self: super: {
-  agent-deck = super.callPackage ../../pkgs/agent-deck.nix { };
-
   ghostship-config = super.writers.writePython3Bin "ghostship-config" {
     libraries = with super.python3Packages; [
       lxml